+++ src/dynamixel/auto_detect.hpp
@@ -59,9 +59,54 @@ namespace dynamixel {
             return std::make_shared<servos::ProM5440S250>(id);
         case servos::ProM4210S260::ct_t::model_number_value:
             return std::make_shared<servos::ProM4210S260>(id);
+        case servos::Mx28P2::ct_t::model_number_value:
+            return std::make_shared<servos::Mx28P2>(id);
+        case servos::Mx64P2::ct_t::model_number_value:
+            return std::make_shared<servos::Mx64P2>(id);
+        case servos::Mx106P2::ct_t::model_number_value:
+            return std::make_shared<servos::Mx106P2>(id);
         default:
             throw errors::Error("Unrecognized model number");
         }
+    }
+
+    /** Send a ping to an ID. If it is answered, read the actuator model and
+        instanciate a class of the correct type.
+
+        @param id value, from 1 to 254 identifiying an actuator
+
+        @return std::shared_ptr to a servo object if we got an answer
+    **/
+    template <typename Protocol, typename Controller>
+    inline std::shared_ptr<servos::BaseServo<Protocol>>
+    find_servo(const Controller& controller, typename Protocol::id_t id)
+    {
+        // Dummy variable used only to differenciate between the two version of
+        // get_servo (protocol 1 or 2)
+        typename Protocol::address_t selected_protocol = 0;
+
+        try {
+            // send a ping.
+            controller.send(instructions::Ping<Protocol>(id));
+            StatusPacket<Protocol> status;
+
+            if (controller.recv(status) && status.id() == id) {
+                // get actuator model
+                controller.send(instructions::Read<Protocol>(id, 0, 2));
+                if (controller.recv(status)) {
+                    uint16_t model;
+                    Protocol::unpack_data(status.parameters(), model);
+                    return get_servo(id, model, selected_protocol);
+                }
+            }
+        }
+        catch (const errors::Error&) {
+            // TODO: This behavior is "dangerous". We should at least tell the
+            //       user something is happening, or give the option for this
+            //       function to re-raise exceptions
+        }
+
+        return std::shared_ptr<servos::BaseServo<Protocol>>();
     }
 
     /** Auto-detect all connected actuators using a given protocol.
@@ -79,35 +124,51 @@ namespace dynamixel {
         @return vector of actuators
     **/
     template <typename Protocol, typename Controller>
-    inline std::vector<std::shared_ptr<servos::BaseServo<Protocol>>> auto_detect(const Controller& controller)
+    inline std::vector<std::shared_ptr<servos::BaseServo<Protocol>>>
+    auto_detect(const Controller& controller)
     {
-        // vector of actuators returned by this function
-        std::vector<std::shared_ptr<servos::BaseServo<Protocol>>> res;
+        using ServoPtr = std::shared_ptr<servos::BaseServo<Protocol>>;
 
-        // Dummy variable used only to differenciate between the tow version of
-        // get_servo (protocol 1 or 2)
-        typename Protocol::address_t selected_protocol = 0;
+        // vector of actuators returned by this function
+        std::vector<ServoPtr> res;
 
         // search through each possible device ID
         for (typename Protocol::id_t id = 0; id < Protocol::broadcast_id; id++) {
-            try {
-                // Send a ping. If it is answered, read the actuator model and
-                // instanciate a class of the correct type
-                controller.send(instructions::Ping<Protocol>(id));
-                StatusPacket<Protocol> status;
-                if (controller.recv(status) && status.id() == id) {
-                    // get actuator model
-                    controller.send(instructions::Read<Protocol>(id, 0, 2));
-                    if (controller.recv(status)) {
-                        uint16_t model;
-                        Protocol::unpack_data(status.parameters(), model);
-                        res.push_back(get_servo(id, model, selected_protocol));
-                    }
-                }
-            }
-            catch (const errors::Error&) {
-                continue;
-            }
+            ServoPtr servo = find_servo<Protocol>(controller, id);
+            if (servo)
+                res.push_back(servo);
+        }
+
+        return res;
+    }
+
+    /** Auto-detect all connected actuators using a given protocol
+
+        @see template <typename Protocol, typename Controller>
+            inline std::vector<std::shared_ptr<servos::BaseServo<Protocol>>>
+            auto_detect(const Controller& controller)
+
+        @param controller object handling the USB to dynamixel interface, instance
+            of the dynamixel::controllers::Usb2Dynamixel class
+        @param ids vector of ids (1-254) that will be searched
+        @return vector of actuators
+    **/
+    template <typename Protocol, typename Controller>
+    inline std::vector<std::shared_ptr<servos::BaseServo<Protocol>>>
+    auto_detect(
+        const Controller& controller,
+        const std::vector<typename Protocol::id_t>& ids)
+    {
+        using ServoPtr = std::shared_ptr<servos::BaseServo<Protocol>>;
+
+        // vector of actuators returned by this function
+        std::vector<ServoPtr> res;
+
+        // search through each possible device ID
+        for (typename Protocol::id_t id : ids) {
+            ServoPtr servo = find_servo<Protocol>(controller, id);
+            if (servo)
+                res.push_back(servo);
         }
 
         return res;
@@ -131,38 +192,54 @@ namespace dynamixel {
     inline std::map<typename Protocol::id_t, std::shared_ptr<servos::BaseServo<Protocol>>>
     auto_detect_map(const Controller& controller)
     {
-        // vector of actuators returned by this function
-        std::map<typename Protocol::id_t, std::shared_ptr<servos::BaseServo<Protocol>>>
-            res;
+        using ServoPtr = std::shared_ptr<servos::BaseServo<Protocol>>;
 
-        // Dummy variable used only to differenciate between the two versions of
-        // get_servo (protocol 1 or 2)
-        typename Protocol::address_t selected_protocol = 0;
+        // vector of actuators returned by this function
+        std::map<typename Protocol::id_t, ServoPtr> res;
 
         // search through each possible device ID
         for (typename Protocol::id_t id = 0; id < Protocol::broadcast_id; id++) {
-            try {
-                // Send a ping. If it is answered, read the actuator model and
-                // instanciate a class of the correct type
-                controller.send(instructions::Ping<Protocol>(id));
-                StatusPacket<Protocol> status;
-                if (controller.recv(status) && status.id() == id) {
-                    // get actuator model
-                    controller.send(instructions::Read<Protocol>(id, 0, 2));
-                    if (controller.recv(status)) {
-                        uint16_t model;
-                        Protocol::unpack_data(status.parameters(), model);
-                        res[id] = (get_servo(id, model, selected_protocol));
-                    }
-                }
-            }
-            catch (const errors::Error&) {
-                continue;
-            }
+            ServoPtr servo = find_servo<Protocol>(controller, id);
+            if (servo)
+                res[id] = servo;
         }
 
         return res;
     }
-}
+
+    /** Auto-detect all connected actuators using a given protocol.
+
+        @see template <typename Protocol, typename Controller>
+            inline std::map<typename Protocol::id_t, std::shared_ptr<servos::BaseServo<Protocol>>>
+            auto_detect_map(const Controller& controller)
+
+        @param controller object handling the USB to dynamixel interface, instance
+            of the dynamixel::controllers::Usb2Dynamixel class
+        @return vector of actuators
+        @throws dynamixel::errors::Errors if there is a problem during send
+        @throws dynamixel::errors::UnpackError if the size of received packet is
+            not correct
+    **/
+    template <typename Protocol, typename Controller>
+    inline std::map<typename Protocol::id_t, std::shared_ptr<servos::BaseServo<Protocol>>>
+    auto_detect_map(
+        const Controller& controller,
+        const std::vector<typename Protocol::id_t>& ids)
+    {
+        using ServoPtr = std::shared_ptr<servos::BaseServo<Protocol>>;
+
+        // vector of actuators returned by this function
+        std::map<typename Protocol::id_t, ServoPtr> res;
+
+        // search through each possible device ID
+        for (typename Protocol::id_t id : ids) {
+            ServoPtr servo = find_servo<Protocol>(controller, id);
+            if (servo)
+                res[id] = servo;
+        }
+
+        return res;
+    }
+} // namespace dynamixel
 
 #endif