+++ src/dynamixel/controllers/usb2dynamixel.hpp
@@ -19,6 +19,57 @@
 #include "../status_packet.hpp"
 
 namespace dynamixel {
+    /** Give an explanation for the error number associated to `write`.
+
+        @param error number, usually the value of errno
+        @return std::string of the explanation
+    **/
+    std::string write_error_string(int error_number)
+    {
+        switch (error_number) {
+        case EAGAIN:
+            return "EAGAIN: The file descriptor fd refers to a file other than "
+                   "a socket and has been marked nonblocking (O_NONBLOCK), and "
+                   "the write would block.";
+        // On some OS, EWOULDBLOCK has the same value as EAGAIN and would hence not compile.
+        // case EWOULDBLOCK:
+        //     return "EWOULDBLOCK: The file descriptor fd refers to a socket "
+        //            "and has been marked nonblocking (O_NONBLOCK), and the "
+        //            "write would block.";
+        case EBADF:
+            return "EBADF: fd is not a valid file descriptor or is not open for writing.";
+        case EDESTADDRREQ:
+            return "EDESTADDRREQ: fd refers to a datagram socket for which a "
+                   "peer address has not been set using connect(2).";
+        case EDQUOT:
+            return "EDQUOT: The user's quota of disk blocks on the file system "
+                   "containing the file referred to by fd has been exhausted.";
+        case EFAULT:
+            return "EFAULT: buf is outside your accessible address space.";
+        case EFBIG:
+            return "EFBIG: An attempt was made to write a file that exceeds the "
+                   "implementation-defined maximum file size or the process's "
+                   "file size limit, or to write at a position past the maximum "
+                   "allowed offset.";
+        case EINTR:
+            return "EINTR: The call was interrupted by a signal before any data "
+                   "was written; see signal(7).";
+        case EINVAL:
+            return "EINVAL: fd is attached to an object which is unsuitable for "
+                   "writing; or the file was opened with the O_DIRECT flag, and "
+                   "either the address specified in buf, the value specified in "
+                   "count, or the current file offset is not suitably aligned.";
+        case EIO:
+            return "EIO: A low-level I/O error occurred while modifying the inode.";
+        case ENOSPC:
+            return "ENOSPC: The device containing the file referred to by fd "
+                   "has no room for the data.";
+        case EPIPE:
+            return "EPIPE: fd is connected to a pipe or socket whose reading end is closed.";
+        }
+        return "";
+    }
+
     namespace controllers {
         class Usb2Dynamixel {
             // TODO : declare private copy constructor and assignment operator
@@ -38,6 +89,7 @@ namespace dynamixel {
                 : _recv_timeout(recv_timeout), _fd(-1), _report_bad_packet(false)
             {
                 open_serial(name, baudrate);
+
             }
             Usb2Dynamixel()
                 : _recv_timeout(0.1), _fd(-1), _report_bad_packet(false) {}
@@ -85,15 +137,18 @@ namespace dynamixel {
             **/
             void open_serial(const std::string& name, int baudrate = B115200)
             {
+                std::cout<<"opening serial"<<std::endl;
                 struct termios tio_serial;
 
                 // Firstly, check that there is no active connexion
                 if (_fd != -1)
-                    throw errors::Error("error attempting to open device " + name + ": an other connexion is active; call `close serial` before opening a new connexion");
+                    throw errors::Error("error attempting to open device " + name + ": an other connection is active; call `close serial` before opening a new connection");
 
-                _fd = open(name.c_str(), O_RDWR | O_NOCTTY);
+                _fd = open(name.c_str(), O_RDWR | O_NOCTTY | O_NONBLOCK );
                 if (_fd == -1)
                     throw errors::Error("error opening device " + name + ": " + std::string(strerror(errno)));
+
+                std::cout<<"serial open"<<std::endl;
 
                 // Serial port setting
                 bzero(&tio_serial, sizeof(tio_serial));
@@ -111,12 +166,20 @@ namespace dynamixel {
                 cfgetispeed(&tio_serial);
                 tcflush(_fd, TCIFLUSH);
                 tcsetattr(_fd, TCSANOW, &tio_serial);
+                std::cout<<"serial ready"<<std::endl;
+
             }
 
             /** Close the serial port
             **/
             void close_serial()
             {
+                // apparently the mac does not flush everything
+                // before closing the fd (Linux does)
+                //... which can cause the last packet to be ignored
+            #ifdef __APPLE__
+                usleep(1000);
+            #endif
                 close(_fd);
                 _fd = -1;
             }
@@ -136,18 +199,21 @@ namespace dynamixel {
                 if (_fd == -1)
                     return;
 
-                int ret = write(_fd, packet.data(), packet.size());
+                const int ret = write(_fd, packet.data(), packet.size());
 
-                /*std::cout << "Send: ";
-            for (size_t i = 0; i < packet.size(); ++i)
-                std::cout << "0x" << std::setfill('0') << std::setw(2) << std::hex << (unsigned int)packet[i] << " ";
-            std::cout << std::endl;*/
+            //     std::cout << "Send: ";
+            // for (size_t i = 0; i < packet.size(); ++i)
+            //     std::cout << "0x" << std::setfill('0') << std::setw(2) << std::hex << (unsigned int)packet[i] << " ";
+            // std::cout << std::endl;
 
-                if (ret != packet.size()) {
+                if (ret == -1) {
+                    throw errors::Error("Usb2Dynamixel::Send write error " + write_error_string(errno));
+                }
+                else if (ret != packet.size()) {
                     std::stringstream ofs;
                     perror("write:");
                     ofs << "written= " << ret << " size=" << packet.size();
-                    throw errors::Error("Usb2Dynamixel::Send: packet not fully written" + ofs.str());
+                    throw errors::Error("Usb2Dynamixel::Send: packet not fully written " + ofs.str());
                 }
             }
 
@@ -166,36 +232,37 @@ namespace dynamixel {
                 std::vector<uint8_t> packet;
                 packet.reserve(_recv_buffer_size);
 
-                //std::cout << "Receive:" << std::endl;
-
+                std::cout << "Receive timeout" << _recv_timeout << std::endl;
                 do {
                     double current_time = get_time();
                     uint8_t byte;
                     int res = read(_fd, &byte, 1);
                     if (res > 0) {
-                        // std::cout << std::setfill('0') << std::setw(2)
-                        //           << std::hex << (unsigned int)byte << " ";
+                         std::cout << std::setfill('0') << std::setw(2)
+                                   << std::hex << (unsigned int)byte << " ";
                         packet.push_back(byte);
 
                         state = status.decode_packet(packet, _report_bad_packet);
+                        status.print(std::cout);
+                        std::cout<<"----"<<std::endl;
                         if (state == DecodeState::INVALID) {
                             // std::cout << "\tBad packet: ";
                             // for (const auto byte : packet)
                             //     std::cout << std::setfill('0') << std::setw(2) << std::hex << (unsigned int)byte << " ";
                             // std::cout << std::endl;
 
-                            packet.clear();
+                            //packet.clear();
                         }
 
                         time = current_time;
                     }
 
-                    if (current_time - time > _recv_timeout)
+                    if (current_time - time > 0.25)//_recv_timeout)
                         return false;
                 } while (state != DecodeState::DONE);
 
                 //std::cout << std::endl;
-                std::cout << std::dec;
+                // std::cout << std::dec;
 
                 return true;
             }
@@ -224,7 +291,7 @@ namespace dynamixel {
             int _fd;
             bool _report_bad_packet;
         };
-    }
-}
+    } // namespace controllers
+} // namespace dynamixel
 
 #endif