+++ src/dynamixel/dynamixel.hpp
@@ -4,5 +4,6 @@
 #include "dynamixel_core.hpp"
 #include "servos.hpp"
 #include "auto_detect.hpp"
+#include "operating_mode.hpp"
 
 #endif