+++ src/dynamixel/protocols/protocol1.hpp
@@ -15,6 +15,9 @@ namespace dynamixel {
     namespace protocols {
         class Protocol1 {
         public:
+            // version of the protocol
+            static constexpr uint8_t version = 1;
+
             typedef uint8_t id_t;
             typedef uint8_t instr_t;
             typedef uint8_t address_t;
@@ -112,15 +115,15 @@ namespace dynamixel {
 
             static void unpack_data(const std::vector<uint8_t>& packet, uint8_t& res)
             {
-                if (packet.size() == 0)
-                    throw errors::UnpackError(1);
+                if (packet.size() != 1)
+                    throw errors::UnpackError(1, packet.size(), 1);
                 res = packet[0];
             }
 
             static void unpack_data(const std::vector<uint8_t>& packet, uint16_t& res)
             {
                 if (packet.size() != 2)
-                    throw errors::UnpackError(1);
+                    throw errors::UnpackError(1, packet.size(), 2);
                 res = (((uint16_t)packet[1]) << 8) | ((uint16_t)packet[0]);
             }
 
@@ -253,7 +256,7 @@ namespace dynamixel {
                 return ~checksum;
             }
         };
-    }
-}
+    } // namespace protocols
+} // namespace dynamixel
 
 #endif